move code to handle the recipe to the multi block class directly
