+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/production/algae/MTEAlgaePondBase.java
@@ -9,6 +9,8 @@ import static gregtech.api.enums.HatchElement.InputHatch;
 import static gregtech.api.enums.HatchElement.OutputBus;
 import static gregtech.api.util.GTStructureUtility.buildHatchAdder;
 
+import java.util.ArrayList;
+import java.util.Random;
 import java.util.stream.Stream;
 
 import javax.annotation.Nonnull;
@@ -30,6 +32,7 @@ import com.gtnewhorizon.structurelib.structure.StructureDefinition;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import gregtech.api.GregTechAPI;
+import gregtech.api.enums.GTValues;
 import gregtech.api.enums.SoundResource;
 import gregtech.api.enums.TAE;
 import gregtech.api.interfaces.IIconContainer;
@@ -37,22 +40,26 @@ import gregtech.api.interfaces.metatileentity.IMetaTileEntity;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.logic.ProcessingLogic;
 import gregtech.api.metatileentity.implementations.MTEHatchInput;
+import gregtech.api.objects.XSTR;
 import gregtech.api.recipe.RecipeMap;
 import gregtech.api.recipe.check.CheckRecipeResult;
 import gregtech.api.recipe.check.CheckRecipeResultRegistry;
 import gregtech.api.recipe.check.SimpleCheckRecipeResult;
 import gregtech.api.util.GTRecipe;
 import gregtech.api.util.GTStreamUtil;
+import gregtech.api.util.GTUtility;
 import gregtech.api.util.MultiblockTooltipBuilder;
 import gregtech.api.util.ReflectionUtil;
 import gtPlusPlus.api.objects.Logger;
 import gtPlusPlus.core.block.ModBlocks;
 import gtPlusPlus.core.config.Configuration;
+import gtPlusPlus.core.item.chemistry.AgriculturalChem;
 import gtPlusPlus.core.lib.GTPPCore;
+import gtPlusPlus.core.util.math.MathUtils;
 import gtPlusPlus.core.util.minecraft.FluidUtils;
+import gtPlusPlus.core.util.minecraft.ItemUtils;
 import gtPlusPlus.xmod.gregtech.api.metatileentity.implementations.base.GTPPMultiBlockBase;
 import gtPlusPlus.xmod.gregtech.common.blocks.textures.TexturesGtBlock;
-import gtPlusPlus.xmod.gregtech.loaders.recipe.RecipesAlgaeFarm;
 import ic2.core.init.BlocksItems;
 import ic2.core.init.InternalName;
 import tectech.thing.casing.TTCasingsContainer;
@@ -335,7 +342,7 @@ public class MTEAlgaePondBase extends GTPPMultiBlockBase<MTEAlgaePondBase> imple
             @Nonnull
             @Override
             protected Stream<GTRecipe> findRecipeMatches(@Nullable RecipeMap<?> map) {
-                return GTStreamUtil.ofNullable(RecipesAlgaeFarm.getTieredRecipe(mLevel, inputItems));
+                return GTStreamUtil.ofNullable(getTieredRecipe(mLevel, inputItems));
             }
 
             @NotNull
@@ -383,5 +390,142 @@ public class MTEAlgaePondBase extends GTPPMultiBlockBase<MTEAlgaePondBase> imple
     @Override
     protected SoundResource getActivitySoundLoop() {
         return SoundResource.GT_MACHINES_ALGAE_LOOP;
+    }
+
+    private static GTRecipe getTieredRecipe(int aTier, ItemStack[] aItemInputs) {
+        return generateBaseRecipe(aTier, isUsingCompost(aItemInputs, aTier));
+    }
+
+    private static boolean isUsingCompost(ItemStack[] aItemInputs, int aTier) {
+        ItemStack aCompost = ItemUtils.getSimpleStack(AgriculturalChem.mCompost, 1);
+        for (ItemStack i : aItemInputs) {
+            if (GTUtility.areStacksEqual(aCompost, i)) {
+                if (i.stackSize >= compostForTier(aTier)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    private static int compostForTier(int aTier) {
+        return aTier > 1 ? (int) Math.min(64, Math.pow(2, aTier - 1)) : 1;
+    }
+
+    private static GTRecipe generateBaseRecipe(int aTier, boolean isUsingCompost) {
+
+        if (aTier < 0) return null; // Type Safety
+
+        final ItemStack[] aInputs;
+        if (isUsingCompost) {
+            // Make it use 4 compost per tier if we have some available
+            // Compost consumption maxes out at 1 stack per cycle
+            ItemStack aCompost = ItemUtils.getSimpleStack(AgriculturalChem.mCompost, compostForTier(aTier));
+            aInputs = new ItemStack[] { aCompost };
+            // Boost Tier by one if using compost, so it gets a speed boost
+            aTier++;
+        } else {
+            aInputs = new ItemStack[] {};
+        }
+
+        ItemStack[] aOutputs = getOutputsForTier(aTier);
+        GTRecipe tRecipe = new GTRecipe(
+            false,
+            aInputs,
+            aOutputs,
+            null,
+            new int[] {},
+            new FluidStack[] { GTValues.NF },
+            new FluidStack[] { GTValues.NF },
+            getRecipeDuration(aTier),
+            0,
+            0);
+        tRecipe.mSpecialValue = tRecipe.hashCode();
+        return tRecipe;
+    }
+
+    private static final int[] aDurations = new int[] { 2000, 1800, 1600, 1400, 1200, 1000, 512, 256, 128, 64, 32, 16,
+        8, 4, 2, 1 };
+    private static final Random random = new XSTR();
+
+    private static int getRecipeDuration(int aTier) {
+        final float randFloat = random.nextFloat();
+        float randMult;
+        if (randFloat < 0.96237624) randMult = 1f;
+        else if (randFloat < 0.9912871) randMult = 2f;
+        else randMult = 3f;
+        return (int) (aDurations[aTier] * randMult / 2);
+    }
+
+    private static ItemStack[] getOutputsForTier(int aTier) {
+        ArrayList<ItemStack> outputList = new ArrayList<>();
+
+        if (aTier >= 0) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mAlgaeBiosmass, 2));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mAlgaeBiosmass, 4));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 2));
+            }
+        }
+        if (aTier >= 1) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mAlgaeBiosmass, 4));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 2));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 4));
+            }
+        }
+        if (aTier >= 2) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 2));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 3));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 8));
+            }
+        }
+        if (aTier >= 3) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, 4));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, 1));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, 4));
+            }
+        }
+        if (aTier >= 4) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, 2));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, 3));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGoldenBrownAlgaeBiosmass, 4));
+            }
+        }
+        if (aTier >= 5) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, 4));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGoldenBrownAlgaeBiosmass, 2));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mRedAlgaeBiosmass, 4));
+            }
+        }
+        if (aTier >= 6) {
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGoldenBrownAlgaeBiosmass, 4));
+            outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mRedAlgaeBiosmass, 2));
+            if (MathUtils.randInt(0, 10) > 9) {
+                outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mRedAlgaeBiosmass, 8));
+            }
+            // Iterate a special loop at higher tiers to provide more Red/Gold Algae.
+            for (int i = 0; i < 20; i++) {
+                if (aTier >= (6 + i)) {
+                    int aMulti = i + 1;
+                    outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGreenAlgaeBiosmass, aMulti * 4));
+                    outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mBrownAlgaeBiosmass, aMulti * 3));
+                    outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mGoldenBrownAlgaeBiosmass, aMulti * 2));
+                    outputList.add(ItemUtils.getSimpleStack(AgriculturalChem.mRedAlgaeBiosmass, aMulti));
+                } else {
+                    break;
+                }
+            }
+        }
+
+        final ItemStack[] aOutputs = new ItemStack[outputList.size()];
+        for (int i = 0; i < outputList.size(); i++) {
+            aOutputs[i] = outputList.get(i);
+        }
+        return aOutputs;
     }
 }
